Fix hwloc narrowing warning on 32-bit architectures
hwloc always uses a 64-bit integer for cache size, while
size_t might be 32 bits on a 32-bit architecture. Since we
do not expect any 32-bit architectures to come with more
than 4GB of cache, we simply cast it explicitly to size_t.

+++ src/gromacs/hardware/hardwaretopology.cpp
@@ -301,7 +301,7 @@ parseHwLocCache(const hwloc_topology_t             topo,
 
                 machine->caches.push_back( {
                                                static_cast<int>(cache->attr->cache.depth),
-                                               cache->attr->cache.size,
+                                               static_cast<std::size_t>(cache->attr->cache.size),
                                                static_cast<int>(cache->attr->cache.linesize),
                                                static_cast<int>(cache->attr->cache.associativity),
                                                std::max(static_cast<int>(hwThreads.size()), 1)
